+++ src/util/shuffle.js
